+++ admin/src/content/mod.rs
@@ -1,10 +1,15 @@
-mod content_editor;
+cfg_if::cfg_if! {if #[cfg(feature = "hydrate")] {
+    mod content_editor;
+    mod content_html;
+    mod content_image;
+    mod content_link;
+    pub use content_editor::*;
+    pub use content_html::*;
+    pub use content_image::*;
+    pub use content_link::*;
+}}
+
 mod content_error;
-mod content_html;
-mod content_image;
-mod content_link;
-pub use content_editor::*;
 pub use content_error::*;
-pub use content_html::*;
-pub use content_image::*;
-pub use content_link::*;
+mod content_update;
+pub use content_update::*;
